requirements.txt
pinecone-client==3.0.2
